Fix ratelimit issue in token validation (#35)
* Fix ratelimit issue in token validation

Troi using liblistenbrainz was constantly facing more ratelimit errors than expected,
and on debugging the issue turned out to be in token validation. The auth token is
set on the client after the validate-token endpoint api call, but that means the validate
token endpoint itself is subjected to the lower ip based ratelimit. The validate-token
endpoint recommends and supports passing the token in the authorization header for the
check, and that way the higher ratelimit would be applicable as well. Hence, the change.

* Remove is_token_valid method

To avoid confusion since the token passed is not used anyway and the only way
to test the validity of the token would be by calling set_auth_token in that case.

* update tests